Add files via upload
Generation of entry and target points for the needle, to be used on the dealII software. The points are used to create 64 .vtk files that contain the needle simulations and their resultant deformation, where the simulated tissue is in the form of an unstructured grid. The generation of points considers the next step in the pipeline where data augmentation is used by rotating the grids in the .vtk files 24 times. Therefore the points are generated so that the entry points are from only one quadrant on the Z-Y plane and the target points can be anywhere within the grid volume, thus ensuring that all the rotations batches are distinct.