Importing VTK files and generation of rotation batches.
Since the data is structured data but represented in an unstructured grid, the data arrays for force and displacement magnitudes need  to be extracted and distributed over a new structured grid by interpolation. Rotation batches can then be created by rotating each 24 times, as the data is represented in a cube and a cube has 24 possible rotational variations.