Training, validation and visualization of the output.
After collecting all the rotation batches for both the force and displacement, we obtain a dataset of 1536 samples for each, force and displacement, split into 1392 samples for training and 144 for testing the model. The model is a 3DUNET and the implementation utilizes its autoencoding functionality, where the input data is the force magnitude which is mapped into the corresponding displacement. Training and Testing losses are computed after every epoch.